Added output test file for scrutiny
These are the text files containing the proposed wikitext that the bot would have saved on the corresponding pages for each file. Inside is what would be the code for the file (but ignoring the Bulgarian-entry check, i.e. attempting to place new derived form entries regardless of whether there already existed one or not.