Uploaded test outputs
These are the files that represent the only changes the bot would have actually committed if let loose 'in the wild,' as the vast majority of Bulgarian entries near the top of the category already have definitions of some kind. These are the ones without any existing Bulgarian definition. Hopefully there are no problems to see here.